+++ NEWS.md
@@ -4,6 +4,8 @@
 
 ## New features
 
+<<<<<<< HEAD
+* Creating default print.sr_model function for sr_model class (#114)
 * Including fitted and residual values as data frame in run_mod output. (#101)
 * Added  `plot_predicted_curve()` with support for faceting by multiple IDs (#68)
 * Replacing old data object with new run_mod output (#102)